worked on functionality of buttons, running but not working

+++ app/src/main/java/com/example/rezepte/adapters/HauptgerichteAdapter.kt
@@ -42,7 +42,7 @@ class HauptgerichteAdapter(private val data: List<Rezept>) : RecyclerView.Adapte
 
         override fun onClick(v: View?) {
             val intent: Intent = Intent(context, RezeptDetailFragment::class.java)
-            var rezept_id: Long = data[adapterPosition].id
+            var rezept_id: Int = data[adapterPosition].id
             var rezept_titel: String = data[adapterPosition].Titel
             var rezept_zutaten: String = data[adapterPosition].Zutaten
             var rezept_beschreibung: String = data[adapterPosition].Beschreibung
